Moved search to home screen
